hw/mem/system-memory: add a memory sysbus device
This device can be used to create some memories using standard
device_add qmp command.

This device has one property 'readonly' which allows
to choose between a ram or a rom.
The device holds the adequate memory region and can be put on
the sysbus.

Signed-off-by: Damien Hedde <damien.hedde@greensocs.com>
---

Should we add a related CONFIG_ variable in the build system ?

Depending on the chosen condition to add a device, the commit may
change.