qdev-monitor: prevent conflicts between qmp/device_add and cli/-device
This commit prepares to extend device_add qmp command when using
-preconfig option.

In order to avoid conflicts with the cli -device option handling, we
need to handle some special case with the QemuOpts.
The qemu_device_opts is traversed when switching from
MACHINE_INIT_PHASE_INITIALIZED to MACHINE_INIT_PHASE_READY in order
to create any device specified by cli -device. Until now any
device_add qmp command was issued after that point so there was no
problem.

If we execute the qmp command before the MACHINE_INIT_PHASE_READY
phase we need to discard the QemuOpts from the qemu_device_opts in
order to avoid the cli -device code to try to create the device
again.

This commit preserves the opts behavior regarding the devices added
in 'ready' phase by the QMP command device_add.

Signed-off-by: Damien Hedde <damien.hedde@greensocs.com>
---

Although we keep the original behavior for QMP commands issued when
the machine is ready (only authorized case so far), we are not sure
it is necessary: keeping the opts in the list is not needed anymore
to ensure the id uniqueness of devices but it has the 2 following
consequences:

1. the device opts stay in the QemuOptsList. Is this list needed
   after traversing the device cli options?

2. the DeviceState "opts" field is set. Do we need to keep it after
   the device is realized ?

Any information on this will be appreciated.