docs/devel: add doc about device life cycles
Document the 3 life cycles case that can happen with devices.

Signed-off-by: Damien Hedde <damien.hedde@greensocs.com>
---

Hi all,

It's been a few weeks I wanted to propose this in order to sort
out what should be done to make a 'user-creatable' device.

This is a follow up of [1] in which Peter asked for this point to
be clarified.

[1]: https://lore.kernel.org/qemu-devel/a2967493-fd00-8f9b-29bd-56baaae9f89a@greensocs.com/

Thanks,
Damien